Moving files into a better folder structure.
