+++ src/exceptions/GameOverException.java
@@ -1,4 +1,4 @@
-package logic.game;
+package exceptions;
 
 public class GameOverException extends Exception {    
 }
