+++ src/services/MineRevealService.java
@@ -1,6 +1,9 @@
-package logic.game;
+package services;
 
+import exceptions.GameOverException;
 import java.util.ArrayList;
+import models.Mine;
+import models.Mines;
 
 public class MineRevealService {
     private OctoCheckService octoCheckService;
